Fixed Kotlin test and replaced a few implicit castings with static_cast calls

+++ src/Bravo/Asset.cpp
@@ -51,8 +51,8 @@ Asset Asset::fromString(std::string assetString) {
         throw std::invalid_argument("Missing decimal fraction after decimal point");
     }
 
-    uint8_t decimals = (dotPosition == string::npos) ? 0 : amountString.size() - dotPosition - 1;
-    int64_t precision = pow(10, static_cast<double>(decimals));
+    uint8_t decimals = (dotPosition == string::npos) ? 0 : static_cast<uint8_t>(amountString.size() - dotPosition - 1);
+    int64_t precision = static_cast<uint64_t>(pow(10, static_cast<double>(decimals)));
 
     // Parse amount
     int64_t intPart, fractPart = 0;
